Project structure adjustments in compliance with existing convention.
