+++ bundles/pubsub/pubsub_protocol/pubsub_protocol_lib/gtest/src/PS_WP_common_ei_tests.cc
@@ -22,8 +22,8 @@
 #include <gtest/gtest.h>
 #include <iostream>
 #include <cstring>
-#include <celix_properties_ei.h>
-#include <malloc_ei.h>
+#include "celix_properties_ei.h"
+#include "malloc_ei.h"
 
 #include "pubsub_wire_protocol_common.h"
