+++ libs/error_injector/api/celix_error_injector.h
@@ -30,7 +30,7 @@ extern "C" {
 #define CELIX_EI_GET_CALLER(addr, level)                                   \
 do {                                                                       \
     Dl_info dlinfo;                                                        \
-    dladdr(__builtin_return_address(level), &dlinfo);                     \
+    dladdr(__builtin_return_address(level), &dlinfo);                      \
     (addr) = dlinfo.dli_saddr;                                             \
 } while(0)
